fix: noise_floor/battery_mv/uptime_secs scanned as float64 to handle REAL values
SQLite stores these as REAL on some instances. Go *int scan silently
fails, dropping the entire observer row (404 on detail, missing from list).
Reported for YC-Base-Repeater and YC-Work-Repeater.

+++ cmd/server/db.go
@@ -142,9 +142,9 @@ type Observer struct {
 	Firmware      *string `json:"firmware"`
 	ClientVersion *string `json:"client_version"`
 	Radio         *string `json:"radio"`
-	BatteryMv     *int    `json:"battery_mv"`
-	UptimeSecs    *int    `json:"uptime_secs"`
-	NoiseFloor    *int    `json:"noise_floor"`
+	BatteryMv     *float64 `json:"battery_mv"`
+	UptimeSecs    *float64 `json:"uptime_secs"`
+	NoiseFloor    *float64 `json:"noise_floor"`
 }
 
 // Transmission represents a row from the transmissions table.
